fix #15, Adding a couple of fallback routes for when getting the AssemblyName from the embedded dll.

+++ src/Jellyfin.Plugin.FileTransformation/ModuleInitializer.cs
@@ -26,21 +26,68 @@ namespace Jellyfin.Plugin.FileTransformation
                 assemblyStream!.CopyTo(memoryStream);
                 assemblyStream.Position = 0;
                 
-                string tmpDllLocation = $"{Path.GetTempFileName()}.dll";
-
+                string? tmpDllLocation = $"{Path.GetTempFileName()}.dll";
+                AssemblyName? assemblyName = null;
+                
                 if (applicationPaths != null)
                 {
                     tmpDllLocation = Path.Combine(applicationPaths.TempDirectory, Path.GetFileName(tmpDllLocation));
+                    
+                    FileInfo dllFileInfo = new FileInfo(tmpDllLocation);
+
+                    try
+                    {
+                        dllFileInfo.Directory?.Create();
+                    }
+                    catch
+                    {
+                        // We tried the supplied jellyfin TempDirectory and we didn't have write permissions. We're going to just use the configurations directory
+                        // as if this fails the user has other problems.
+                        logger?.LogWarning($"Unable to create temp directory for file transformation plugin. Tried '{applicationPaths.TempDirectory}'. Now attempting '{applicationPaths.ConfigurationDirectoryPath}' before failing.");
+
+                        try
+                        {
+                            tmpDllLocation = Path.Combine(applicationPaths.ConfigurationDirectoryPath, Path.GetFileName(tmpDllLocation));
+                            dllFileInfo = new FileInfo(tmpDllLocation);
+                            dllFileInfo.Directory?.Create();
+                        }
+                        catch
+                        {
+                            logger?.LogError($"Unable to create temp directory for file transformation plugin. Tried '{applicationPaths.TempDirectory}' and '{applicationPaths.ConfigurationDirectoryPath}' as directories, neither allowed writing. Falling back to using a fallback AssemblyLoadContext and loading DLL to get assembly name before unloading and loading properly.");
+                            tmpDllLocation = null;
+                        }
+                    }
                 }
-                
-                logger?.LogInformation($"Writing dll to: {tmpDllLocation} to extract AssemblyName details, will be removed after loading");
-                File.WriteAllBytes(tmpDllLocation, memoryStream.ToArray());
-                
-                AssemblyName assemblyName = AssemblyName.GetAssemblyName(tmpDllLocation);
-                
-                logger?.LogInformation($"Deleting: {tmpDllLocation}");
-                File.Delete(tmpDllLocation);
-                
+
+                try
+                {
+                    if (tmpDllLocation == null)
+                    {
+                        // We're expecting this when we can't create the directories. Lets just send an exception to enter the catch
+                        throw new FileNotFoundException($"Null DLL location.");
+                    }
+                    
+                    logger?.LogInformation($"Writing dll to: {tmpDllLocation} to extract AssemblyName details, will be removed after loading");
+                    File.WriteAllBytes(tmpDllLocation, memoryStream.ToArray());
+
+                    assemblyName = AssemblyName.GetAssemblyName(tmpDllLocation);
+
+                    logger?.LogInformation($"Deleting: {tmpDllLocation}");
+                    File.Delete(tmpDllLocation);
+                }
+                catch
+                {
+                    // Attempting final fallback
+                    AssemblyLoadContext throwAwayContext = new AssemblyLoadContext($"{typeof(ModuleInitializer).Namespace}.FallbackContext", true);
+                    Assembly tmpLoadedAssembly = throwAwayContext.LoadFromStream(assemblyStream);
+                    assemblyStream.Position = 0;
+
+                    assemblyName = tmpLoadedAssembly.GetName();
+                            
+                    logger?.LogInformation($"Retrieved AssemblyName from fallback context '{throwAwayContext.Name}'. Now unloading");
+                    throwAwayContext.Unload();
+                }
+
                 Assembly loadedAssembly;
                 if (!assemblyLoadContext.Assemblies.Any(x => x.FullName == assemblyName.FullName))
                 {
